+++ include/nstd/core/str.h
@@ -1,7 +1,7 @@
 #ifndef NSTD_CORE_STR_H
 #define NSTD_CORE_STR_H
 #include "../nstd.h"
-#include "cstr.h"
+#include "cstr/cstr.h"
 #include "def.h"
 #include "optional.h"
 #include "range.h"
@@ -19,6 +19,9 @@ typedef struct {
     /// The number of bytes in the string.
     NSTDUInt len;
 } NSTDStr;
+
+/// Represents an optional value of type `NSTDStr`.
+NSTDOptional(NSTDStr) NSTDOptionalStr;
 
 /// Creates a new instance of an `NSTDStr` from a C string slice.
 ///
@@ -73,6 +76,8 @@ NSTDAPI NSTDStr nstd_core_str_from_cstr_unchecked(const NSTDCStr *cstr);
 ///
 /// This function will panic in the following situations:
 ///
+/// - `cstr` is null.
+///
 /// - `cstr`'s data is not valid UTF-8.
 ///
 /// - `cstr`'s length is greater than `NSTDInt`'s max value.
@@ -96,6 +101,8 @@ NSTDAPI NSTDStr nstd_core_str_from_raw_cstr(const NSTDChar *cstr);
 /// # Panics
 ///
 /// This function will panic in the following situations:
+///
+/// - `cstr` is null.
 ///
 /// - `cstr`'s data is not valid UTF-8.
 ///
@@ -156,6 +163,17 @@ NSTDAPI NSTDStr nstd_core_str_from_bytes(const NSTDSlice *bytes);
 ///
 /// - `bytes`'s data must be valid for reads of at least `bytes.len` consecutive bytes.
 NSTDAPI NSTDStr nstd_core_str_from_bytes_unchecked(const NSTDSlice *bytes);
+
+/// Returns a C string slice variant of this UTF-8 encoded string slice.
+///
+/// # Parameters:
+///
+/// - `const NSTDStr *str` - The UTF-8 encoded string slice.
+///
+/// # Returns
+///
+/// `NSTDCStr cstr` - The new C string slice.
+NSTDAPI NSTDCStr nstd_core_str_as_cstr(const NSTDStr *str);
 
 /// Returns an immutable byte slice over `str`'s data.
 ///
@@ -511,6 +529,9 @@ typedef struct {
     NSTDUInt len;
 } NSTDStrMut;
 
+/// Represents an optional value of type `NSTDStrMut`.
+NSTDOptional(NSTDStrMut) NSTDOptionalStrMut;
+
 /// Creates a new instance of an `NSTDStrMut` from a C string slice.
 ///
 /// # Parameters:
@@ -564,6 +585,8 @@ NSTDAPI NSTDStrMut nstd_core_str_mut_from_cstr_unchecked(NSTDCStrMut *cstr);
 ///
 /// This function will panic in the following situations:
 ///
+/// - `cstr` is null.
+///
 /// - `cstr`'s data is not valid UTF-8.
 ///
 /// - `cstr`'s length is greater than `NSTDInt`'s max value.
@@ -587,6 +610,8 @@ NSTDAPI NSTDStrMut nstd_core_str_mut_from_raw_cstr(NSTDChar *cstr);
 /// # Panics
 ///
 /// This function will panic in the following situations:
+///
+/// - `cstr` is null.
 ///
 /// - `cstr`'s data is not valid UTF-8.
 ///
@@ -658,6 +683,17 @@ NSTDAPI NSTDStrMut nstd_core_str_mut_from_bytes_unchecked(NSTDSliceMut *bytes);
 ///
 /// `NSTDStr str_const` - The immutable copy of `str`.
 NSTDAPI NSTDStr nstd_core_str_mut_as_const(const NSTDStrMut *str);
+
+/// Returns a C string slice variant of this UTF-8 encoded string slice.
+///
+/// # Parameters:
+///
+/// - `const NSTDStrMut *str` - The UTF-8 encoded string slice.
+///
+/// # Returns
+///
+/// `NSTDCStr cstr` - The new C string slice.
+NSTDAPI NSTDCStr nstd_core_str_mut_as_cstr(const NSTDStrMut *str);
 
 /// Returns an immutable byte slice over `str`'s data.
 ///