+++ include/nstd/vec.h
@@ -353,10 +353,6 @@ NSTDAPI void nstd_vec_truncate(NSTDVec *vec, NSTDUInt len);
 /// # Returns
 ///
 /// `NSTDAllocError errc` - The allocation operation error code.
-///
-/// # Panics
-///
-/// This operation will panic if `size` is zero.
 NSTDAPI NSTDAllocError nstd_vec_reserve(NSTDVec *vec, NSTDUInt size);
 
 /// Decreases a vector's capacity to match it's length.