+++ src/core/ptr.rs
@@ -1,5 +1,12 @@
 //! A sized pointer to some arbitrary type.
-use crate::{core::mem::nstd_core_mem_copy, NSTDAny, NSTDAnyMut, NSTDUInt};
+pub mod raw;
+use crate::{
+    core::{
+        mem::nstd_core_mem_copy,
+        optional::{gen_optional, NSTDOptional},
+    },
+    NSTDAny, NSTDAnyMut, NSTDUInt,
+};
 
 /// A sized immutable pointer to some arbitrary type.
 ///
@@ -15,6 +22,7 @@ pub struct NSTDPtr {
     /// The size of the object being pointed to.
     size: NSTDUInt,
 }
+gen_optional!(NSTDOptionalPtr, NSTDPtr);
 
 /// Creates a new instance of `NSTDPtr`.
 ///
@@ -27,9 +35,14 @@ pub struct NSTDPtr {
 /// # Returns
 ///
 /// `NSTDPtr ptr` - A new instance of `NSTDPtr` that points to `obj`.
+///
+/// # Panics
+///
+/// Panics if `obj` is null.
 #[inline]
 #[cfg_attr(feature = "clib", no_mangle)]
 pub extern "C" fn nstd_core_ptr_new(obj: NSTDAny, size: NSTDUInt) -> NSTDPtr {
+    assert!(!obj.is_null());
     NSTDPtr { raw: obj, size }
 }
 
@@ -104,6 +117,7 @@ pub struct NSTDPtrMut {
     /// The size of the object being pointed to.
     size: NSTDUInt,
 }
+gen_optional!(NSTDOptionalPtrMut, NSTDPtrMut);
 
 /// Creates a new instance of `NSTDPtrMut`.
 ///
@@ -116,9 +130,14 @@ pub struct NSTDPtrMut {
 /// # Returns
 ///
 /// `NSTDPtrMut ptr` - A new instance of `NSTDPtrMut` that points to `obj`.
+///
+/// # Panics
+///
+/// Panics if `obj` is null.
 #[inline]
 #[cfg_attr(feature = "clib", no_mangle)]
 pub extern "C" fn nstd_core_ptr_mut_new(obj: NSTDAnyMut, size: NSTDUInt) -> NSTDPtrMut {
+    assert!(!obj.is_null());
     NSTDPtrMut { raw: obj, size }
 }